feat: ensure AddSignup is called with correct values
