--- src/modules/signup/data/models/add-account.ts
@@ -1,8 +1,0 @@
-export namespace AccountModel {
-  export interface Params {
-    name: string;
-    email: string;
-    password: string;
-    passwordConfirmation: string;
-  }
-}
