+++ src/modules/signup/domain/usecases/add-signup.ts
@@ -1,0 +1,5 @@
+import { SignupModel } from '../../data/models/add-signup';
+
+export interface AddSignup {
+  add(params: SignupModel.Params): Promise<SignupModel.Result>;
+}
